mcp_servers/cashfree/klavis/mcp_servers/pandoc/requirements.txt
mcp==1.11.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pypandoc_binary>=1.15
google-cloud-storage>=2.10.0
python-dotenv>=1.0.0
requests>=2.31.0